+++ packages/@expo/cli/e2e/__tests__/run-test.ts
@@ -88,7 +88,7 @@ it('runs `npx expo run ios --help`', async () => {
         --no-bundler                     Skip starting the Metro bundler
         --scheme [scheme]                Scheme to build
         --binary <path>                  Path to existing .app or .ipa to install.
-        --configuration <configuration>  Xcode configuration to use. Debug or Release. Default: Debug
+        --configuration <configuration>  Xcode configuration to use. Default: Debug
         -d, --device [device]            Device name, UDID, or "generic" for build-only
         -o, --output <path>              Directory to output the built app binary
         -p, --port <port>                Port to start the Metro bundler on. Default: 8081