+++ packages/@expo/cli/src/run/ios/options/resolveOptions.ts
@@ -3,6 +3,7 @@ import { getConfig } from '@expo/config';
 import type { OSType } from '../../../start/platforms/ios/simctl';
 import { isOSType } from '../../../start/platforms/ios/simctl';
 import { resolveBuildCacheProvider } from '../../../utils/build-cache-providers';
+import type { EnvironmentMode } from '../../../utils/nodeEnv';
 import { profile } from '../../../utils/profile';
 import { resolveBundlerPropsAsync } from '../../resolveBundlerProps';
 import type { BuildProps, Options } from '../XcodeBuild.types';
@@ -13,7 +14,8 @@ import { resolveXcodeProject } from './resolveXcodeProject';
 /** Resolve arguments for the `run:ios` command. */
 export async function resolveOptionsAsync(
   projectRoot: string,
-  options: Options
+  options: Options,
+  mode: EnvironmentMode
 ): Promise<BuildProps> {
   const xcodeProject = resolveXcodeProject(projectRoot);
 
@@ -53,10 +55,7 @@ export async function resolveOptionsAsync(
     projectRoot
   );
 
-  // This optimization skips resetting the Metro cache needlessly.
-  // The cache is reset in `../node_modules/react-native/scripts/react-native-xcode.sh` when the
-  // project is running in Debug and built onto a physical device. It seems that this is done because
-  // the script is run from Xcode and unaware of the CLI instance.
+  // Skip native bundling for Debug device builds to avoid resetting Metro's cache.
   const shouldSkipInitialBundling = configuration === 'Debug' && !isSimulator;
 
   return {
@@ -68,6 +67,7 @@ export async function resolveOptionsAsync(
     device,
     osType,
     configuration,
+    mode,
     shouldSkipInitialBundling,
     buildCache: options.buildCache !== false,
     scheme,