+++ packages/@expo/config/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### 💡 Others
 
+- Allow custom Xcode configuration names in the build cache provider types. ([#49555](https://github.com/expo/expo/pull/49555) by [@ramonclaudio](https://github.com/ramonclaudio))
 - Add experimental `tvos` and `macos` platforms gated by `expriments.outOfTreePlatforms` in config ([#46344](https://github.com/expo/expo/pull/46344) by [@kitten](https://github.com/kitten))
 
 ## 57.0.6 - 2026-07-22